ucm: enable hotword modifier on device open
With SOF restoring controls on resume, when WoV is not enabled we see a
resume time regression because the kernel is copying the hotword model
to the firmware. The solution is to only set the model when we are using
the device.

This patch is one part of the solution in that we trigger
the disable sequence on the modifier when close the hotword device. The
second part will be to set a null model on the modifier's disable
sequence.

The only side effect will be non SOF boards is that the model will be
set each time right before device opens.

BUG=b:175897062
TEST=check resume time and verify WoV works

Change-Id: Icd3a0103b5b5e267992ee065200f77b628208741