system: add dsm echo ref flag
Currently we always open our echo ref with the associated playback
stream. This makes sense for devices that need the echo ref running
speaker protection. However, devices that do not have this limitation
that are only doing playback are wasting cycles moving echo data that is
not needed.

Solution is to put a flag in place to differentiate DSM paths. We will
also need to scan all open streams anytime a stream opens or close to
check if we need to open or close a respective echo reference stream.

BUG=b:215318254
TEST=Check if redrix echo ref open on playback and if drobit echo ref
open on playback

Change-Id: Ie37426627a3487f0ef4dc54a93d0594805f0e9d6